+++ src/Mod/Draft/DraftGui.py
@@ -2313,142 +2313,6 @@ class ScaleTaskPanel:
 ##            msg = "TranslateWidget: Can not translate widget: {0} type: {1}\n".format(w.objectName(),w.metaObject().className())
 ##            FreeCAD.Console.PrintMessage(msg)
 
-class ShapeStringTaskPanel:
-    '''A TaskPanel for ShapeString'''
-    oldValueBuffer = False
-
-    def __init__(self):
-        self.form = QtGui.QWidget()
-        self.form.setObjectName("ShapeStringTaskPanel")
-        self.form.setWindowTitle(translate("draft","ShapeString"))
-        layout = QtGui.QVBoxLayout(self.form)
-        uiFile = QtCore.QFile(u":/ui/TaskShapeString.ui")  #this has to change if ui not in Resource file
-        loader = FreeCADGui.UiLoader()
-        self.task = loader.load(uiFile)
-        layout.addWidget(self.task)
-
-        qStart = FreeCAD.Units.Quantity(0.0, FreeCAD.Units.Length)
-        self.task.sbX.setProperty('rawValue',qStart.Value)
-        self.task.sbX.setProperty('unit',qStart.getUserPreferred()[2])
-        self.task.sbY.setProperty('rawValue',qStart.Value)
-        self.task.sbY.setProperty('unit',qStart.getUserPreferred()[2])
-        self.task.sbZ.setProperty('rawValue',qStart.Value)
-        self.task.sbZ.setProperty('unit',qStart.getUserPreferred()[2])
-        self.task.sbHeight.setProperty('rawValue',10.0)
-        self.task.sbHeight.setProperty('unit',qStart.getUserPreferred()[2])
-
-        self.stringText = translate("draft","Default")
-        self.task.leString.setText(self.stringText)
-        self.platWinDialog("Overwrite")
-        self.task.fcFontFile.setFileName(Draft.getParam("FontFile",""))
-        self.fileSpec = Draft.getParam("FontFile","")
-        self.point = FreeCAD.Vector(0.0,0.0,0.0)
-        self.pointPicked = False
-
-        QtCore.QObject.connect(self.task.fcFontFile,QtCore.SIGNAL("fileNameSelected(const QString&)"),self.fileSelect)
-        QtCore.QObject.connect(self.task.pbReset,QtCore.SIGNAL("clicked()"),self.resetPoint)
-        self.point = None
-        self.view = Draft.get3DView()
-        self.call = self.view.addEventCallback("SoEvent",self.action)
-        FreeCAD.Console.PrintMessage(translate("draft", "Pick ShapeString location point:")+"\n")
-
-
-    def fileSelect(self, fn):
-        self.fileSpec = fn
-
-    def resetPoint(self):
-        self.pointPicked = False
-        origin = FreeCAD.Vector(0.0,0.0,0.0)
-        self.setPoint(origin)
-
-    def action(self,arg):
-        """scene event handler"""
-        import DraftTools
-        if arg["Type"] == "SoKeyboardEvent":
-            if arg["Key"] == "ESCAPE":
-                self.reject()
-        elif arg["Type"] == "SoLocation2Event": #mouse movement detection
-            self.point,ctrlPoint,info = DraftTools.getPoint(self.sourceCmd,arg,noTracker=True)
-            if not self.pointPicked:
-                self.setPoint(self.point)
-        elif arg["Type"] == "SoMouseButtonEvent":
-            if (arg["State"] == "DOWN") and (arg["Button"] == "BUTTON1"):
-                self.setPoint(self.point)
-                self.pointPicked = True
-
-    def setPoint(self, point):
-        self.task.sbX.setProperty('rawValue',point.x)
-        self.task.sbY.setProperty('rawValue',point.y)
-        self.task.sbZ.setProperty('rawValue',point.z)
-
-    def createObject(self):
-        """creates object in the current doc"""
-        dquote = '"'
-        if sys.version_info.major < 3: # Python3: no more unicode
-            String  = 'u' + dquote + str(self.task.leString.text().encode('unicode_escape')) + dquote
-        else:
-            String  = dquote + self.task.leString.text() + dquote
-        FFile = dquote + str(self.fileSpec) + dquote
-
-        Size = str(FreeCAD.Units.Quantity(self.task.sbHeight.text()).Value)
-        Tracking = str(0.0)
-        x = FreeCAD.Units.Quantity(self.task.sbX.text()).Value
-        y = FreeCAD.Units.Quantity(self.task.sbY.text()).Value
-        z = FreeCAD.Units.Quantity(self.task.sbZ.text()).Value
-        ssBase = FreeCAD.Vector(x,y,z)
-        # this try block is almost identical to the one in DraftTools
-        try:
-            qr,sup,points,fil = self.sourceCmd.getStrings()
-            FreeCADGui.addModule("Draft")
-            self.sourceCmd.commit(translate("draft","Create ShapeString"),
-    ['ss=Draft.makeShapeString(String='+String+',FontFile='+FFile+',Size='+Size+',Tracking='+Tracking+')',
-                         'plm=FreeCAD.Placement()',
-                         'plm.Base='+DraftVecUtils.toString(ssBase),
-                         'plm.Rotation.Q='+qr,
-                         'ss.Placement=plm',
-                         'ss.Support='+sup,
-                         'Draft.autogroup(ss)'])
-        except Exception as e:
-            FreeCAD.Console.PrintError("Draft_ShapeString: error delaying commit\n")
-
-    def platWinDialog(self, Flag):
-        ParamGroup = FreeCAD.ParamGet("User parameter:BaseApp/Preferences/Dialog")
-        if Flag == "Overwrite":
-            GroupContent = ParamGroup.GetContents()
-            Found = False
-            if GroupContent:
-                for ParamSet in GroupContent:
-                    if ParamSet[1] == "DontUseNativeFontDialog":
-                        Found = True
-                        break
-
-            if Found == False:
-                ParamGroup.SetBool("DontUseNativeFontDialog", True)     #initialize nonexisting one
-
-            param = ParamGroup.GetBool("DontUseNativeFontDialog")
-            ShapeStringTaskPanel.oldValueBuffer = ParamGroup.GetBool("DontUseNativeDialog")
-            ParamGroup.SetBool("DontUseNativeDialog", param)
-
-        elif Flag == "Restore":
-            ParamGroup.SetBool("DontUseNativeDialog", ShapeStringTaskPanel.oldValueBuffer)
-
-    def accept(self):
-        self.createObject();
-        if self.call: self.view.removeEventCallback("SoEvent",self.call)
-        FreeCADGui.ActiveDocument.resetEdit()
-        FreeCADGui.Snapper.off()
-        self.sourceCmd.creator.finish(self.sourceCmd)
-        self.platWinDialog("Restore")
-        return True
-
-    def reject(self):
-        if self.call: self.view.removeEventCallback("SoEvent",self.call)
-        FreeCADGui.ActiveDocument.resetEdit()
-        FreeCADGui.Snapper.off()
-        self.sourceCmd.creator.finish(self.sourceCmd)
-        self.platWinDialog("Restore")
-        return True
-
 if not hasattr(FreeCADGui,"draftToolBar"):
     FreeCADGui.draftToolBar = DraftToolBar()
 #----End of Python Features Definitions----#
